Factor chat display into its own component
